% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_action_workflow_publish_docker_images}
\alias{create_action_workflow_publish_docker_images}
\title{Create Github Actions Workflow for Docker Image Publish}
\usage{
create_action_workflow_publish_docker_images(
  account,
  images,
  rversions,
  save_as = fs::path(".github", "workflows", "publish-docker-images.yml")
)
}
\arguments{
\item{account}{\code{character} dockerhub account for pushing the images.}

\item{images}{names of docker images to create, subset of
\code{c("r-aws-minimal", "r-aws-spatial", "r-cicd-minimal", "r-cicd-spatial")}}

\item{rversions}{\code{character} vector of R versions; necessary for setting up
the matrix for the Github Actions Workflow.}

\item{save_as}{path for storing the yml file; default
to \code{fs::path(".github", "workflows", "publish-docker-images.yml")}.}
}
\value{
Logical indicating whether a write occurred, invisibly.
}
\description{
A github actions workflow for building and pushing a stack of docker images.
}
