% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_dockerfile}
\alias{create_dockerfile}
\title{Create Dockerfile}
\usage{
create_dockerfile(
  name,
  from,
  tag = "4.1.2",
  script,
  save_as = fs::path("dockerfiles", glue::glue("{name}_{tag}.Dockerfile"))
)
}
\arguments{
\item{name}{name of docker image to create, one of
\code{c("r-aws-minimal", "r-aws-spatial", "r-cicd-minimal", "r-cicd-spatial")}}

\item{from}{\code{character} docker image (\emph{without \code{tag}}) to use as basis.}

\item{tag}{\code{character} tag for the docker image, default to \code{"4.1.2"}}

\item{script}{\code{character} name of installation shell script to be
executed in \verb{docker build} process.}

\item{save_as}{path for storing the dockerfile; default
to \code{fs::path("dockerfiles", glue::glue("{name}_{tag}.Dockerfile"))}.}
}
\value{
Logical indicating whether a write occurred, invisibly.
}
\description{
Different dockerfiles are required for the various docker images.
}
