% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsrocker-package.R
\docType{package}
\name{mdsrocker-package}
\alias{mdsrocker}
\alias{mdsrocker-package}
\title{mdsrocker: Docker images for CI/CD and dockerized AWS Solutions at METER Data Science}
\description{
Framework for creating version-stable METER docker images for R based on rocker-project foundation.
}
\author{
\strong{Maintainer}: Christoph Stepper \email{christoph.stepper@metergroup.com}

Authors:
\itemize{
  \item Hendrik Wagenseil \email{hendrik.wagenseil@metergroup.com}
}

}
\keyword{internal}
