% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rocker_dockerfiles}
\alias{rocker_dockerfiles}
\title{Dockerfile Specifications}
\format{
A \code{tibble} with \emph{4} rows and \emph{3} columns:
\itemize{
\item \emph{name}: name of the docker repository, one of
\code{c("r-aws-minimal", "r-aws-spatial", "r-cicd-minimal", "r-cicd-spatial")},
\item \emph{from}: docker image (without \code{tag}) to use as basis,
\item \emph{script}: name of installation shell script to be
executed in \verb{docker build} process.
}
}
\source{
METERGROUP
}
\usage{
rocker_dockerfiles
}
\description{
A data set containing the the specifications for different dockerfiles.
}
\keyword{datasets}
