% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rocker_installation}
\alias{rocker_installation}
\title{Installation Instructions for Docker Builds}
\format{
A \code{tibble} with \emph{3} rows and \emph{3} columns:
\itemize{
\item \emph{type}: type of software bundle to install, one of
\code{c("aws", "cicd", "spatial")},
\item \emph{pkgs}: R pacakges to install in respective install scripts,
\item \emph{syslibs}: additionally system libraries
(not required by specified R packages).
}
}
\source{
METERGROUP
}
\usage{
rocker_installation
}
\description{
A data set containing the the requirements to install in the docker builds.
}
\keyword{datasets}
