% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_shellscript}
\alias{create_shellscript}
\title{Create Shell Script with Installation Instructions}
\usage{
create_shellscript(
  type,
  rpkgs,
  syslibs = NULL,
  pypkgs = NULL,
  extra = NULL,
  os = "ubuntu-20.04",
  save_as = fs::path("scripts", glue::glue("install_{type}.sh"))
)
}
\arguments{
\item{type}{type of software bundle to install, one of
\code{c("aws", "cicd", "spatial", "full")}}

\item{rpkgs}{\code{character} vector of R packages to install}

\item{syslibs}{\code{character} vector of system libraries to install
(which are not found as dependency of the \code{rpkgs}).}

\item{pypkgs}{\code{character} vector of Python packages to install}

\item{extra}{\code{character} vector of additional software packages to install}

\item{os}{definition of operating system; default to \code{"ubuntu-20.04"}.}

\item{save_as}{path for storing the installation instruction file; default
to \code{fs::path("scripts", glue::glue("install-{type}.sh"))}.}
}
\value{
Logical indicating whether a write occurred, invisibly.
}
\description{
Different shell scripts with installation instructions are required
for the various docker images.
}
