% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_shellscript}
\alias{create_shellscript}
\title{Create Shell Script with Installation Instructions}
\usage{
create_shellscript(
  type,
  pkgs,
  os = "ubuntu-20.04",
  syslibs = NULL,
  save_as = fs::path("scripts", glue::glue("install-{type}.sh"))
)
}
\arguments{
\item{type}{type of software bundle to install, one of
\code{c("aws", "cicd", "spatial")}}

\item{pkgs}{\code{character} vector of R packages to install}

\item{os}{definition of operating system; default to \code{"ubuntu-20.04"}.}

\item{syslibs}{\code{character} vector of system libraries to install
(which are not found as dependency of the R \code{pkgs}.}

\item{save_as}{path for storing the installation instruction file; default
to \code{fs::path("scripts", glue::glue("install-{type}.sh"))}.}
}
\value{
Logical indicating whether a write occurred, invisibly.
}
\description{
Different shell scripts with installation instructions are required
for the various docker images.
}
